xar: Remove unneeded cast
The variable is size_t and argument expects a size_t. Do not cast to
int, which could only lead to trouble (in theory).

+++ libarchive/archive_read_support_format_xar.c
@@ -3638,7 +3638,7 @@ xmllite_read_toc(struct archive_read *a)
 				goto out;
 			}
 
-			xml_data(a, as.s, (int)archive_strlen(&as));
+			xml_data(a, as.s, archive_strlen(&as));
 			archive_string_free(&as);
 
 		case XmlNodeType_None:
